non-seasonal-scenario-planner-api/requirements.txt
azure-core==1.23.0
azure-storage-blob==12.10.0b4
certifi==2018.8.24
cffi==1.15.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
coveralls==1.5.1
cryptography==36.0.1
Django==2.2.10
django-cors-headers
django-cors-middleware
django-extensions==1.7.1
django-filter==2.0.0
django-multiselectfield==0.1.12
django-rest-swagger==2.2.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
et-xmlfile==1.1.0
factory-boy==2.11.1
Faker==0.9.2
idna==2.7
isodate==0.6.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.0.1
MarkupSafe==2.0.1
msrest==0.6.21
mssql-django==1.1.1
numpy
oauthlib==3.1.1
openapi-codec==1.3.2
openpyxl==3.0.9
pandas==1.3.2
psycopg2==2.8.6
pycparser==2.21
PyJWT==1.6.4
pyodbc==4.0.32
python-dateutil==2.8.2
pytz==2021.3
requests==2.20.0
requests-oauthlib==1.3.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.2
uritemplate==4.1.1
urllib3==1.24.2
whitenoise==5.3.0
XlsxWriter==3.0.1
pyspark
fastparquet
azure.identity
azure-storage-file-datalake
azure-storage-file-datalake --pre
wkhtmltopdf
django-wkhtmltopdf
pdfkit